Implement Edit Action in MemeDetailViewController
Open MemeEditorViewController in Edit Mode. Change the default text for
top and bottom textfields to top and bottom text of the meme image
respectively. This clears the textfields when user taps inside it.

In edit mode, saving the meme updates the existing meme in the shared
model.

Adjusted auto layout constraints of top and bottom textfields for
better alignment when keyboard is active.