Refactor MemeEditorViewController
There are other ways to stop editing a textfield apart from the return
key on the keyboard, therefore textFieldDidEndEditing handles text
reset instead of textFieldShouldReturn.

Renamed the notification subscription methods to be more generic since
it manages both keyboard and device orientation change notifications.

With the introduction of custom keyboards in iOS,
UIKeyboardWillShowNotification can be called multiple times by the
keyboard resulting in the view to slide up higher than it should. The
updated code handles this situation.